% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-shadowpath.r
\name{geom_shadowpath}
\alias{geom_shadowpath}
\title{Connect observations, plot a shadow beneath the connected lines to make it easier to read
a chart with several overlapping observations.}
\usage{
geom_shadowpath(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linemitre}{Line mitre limit (number greater than 1).}

\item{arrow}{Arrow specification, as created by [grid::arrow()].}
}
\description{
`geom_shadowpath()` connects the observations in the order in which they appear
in the data. `geom_shadowline()` connects them in order of the variable on the
x axis. `geom_shadowstep()` creates a stairstep plot, highlighting exactly
when changes occur. The `group` aesthetic determines which cases are
connected together. These functions are designed as a straight replacement
to the [geom_path()], [geom_line()] and [geom_step()] functions.
}
\details{
To set the order of drawing, make the `colour` aesthetic a factor, and set the order
from bottom to top.
}
\section{Missing value handling}{

`geom_shadowpath()`, `geom_shadowline()`, and `geom_shadowstep` handle `NA` as follows:

* If an `NA` occurs in the middle of a line, it breaks the line. No warning
  is shown, regardless of whether `na.rm` is `TRUE` or `FALSE`.
* If an `NA` occurs at the start or the end of the line and `na.rm` is `FALSE`
  (default), the `NA` is removed with a warning.
* If an `NA` occurs at the start or the end of the line and `na.rm` is `TRUE`,
  the `NA` is removed silently, without warning.
}

\examples{
# geom_shadowline() is suitable for time series
library(ggplot2)
ggplot(economics_long, aes(date, value01, colour = variable)) +
  geom_shadowline()

}
\seealso{
[geom_path()], [geom_line()], [geom_step()]: Filled paths (polygons);
}
