% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-shadowpoint.r
\name{geom_shadowpoint}
\alias{geom_shadowpoint}
\title{Points}
\usage{
geom_shadowpoint(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}
}
\description{
The point geom is used to create scatterplots. [geom_shadowpoint()] is
designed as a drop in replacement for [geom_point()] with an added shadow beneath
the point to make a busy plot more aesthetically appealing or to make points
stand out from the rest of the plot.
}
\section{Aesthetics}{

Adds 3 new aesthetics to [geom_point()]:
* \code{shadowcolour} defaults to white, controls the color of the shadow.
* \code{shadowsize} defaults to \code{1.8 * size}, controls the sie of the shadow.
* \code{shadowalpha} defaults to \code{0.25 * alpha} or \code{0.9}, controls the alpha of the shadow.
}

\examples{
library( ggplot2 )
p <- ggplot(mtcars, aes(wt, mpg))
p + geom_shadowpoint()

}
