% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-glowpath.r, R/geom-glowpoint.r,
%   R/geom-shadowpath.r, R/geom-shadowpoint.r, R/internal-doc.r
\docType{data}
\name{GeomGlowPath}
\alias{GeomGlowPath}
\alias{GeomGlowLine}
\alias{GeomGlowStep}
\alias{GeomGlowPoint}
\alias{GeomShadowPath}
\alias{GeomShadowLine}
\alias{GeomShadowStep}
\alias{GeomShadowPoint}
\alias{ggshadow-ggproto}
\title{ggproto classes for ggshadow}
\description{
Please refer to ggplot2 documentation for more information on the ggproto class system
}
\keyword{datasets}
\keyword{internal}
