% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshadow-package.R
\docType{package}
\name{ggshadow-package}
\alias{ggshadow}
\alias{ggshadow-package}
\title{ggshadow: Shadow and Glow Geoms for 'ggplot2'}
\description{
A collection of Geoms for R's 'ggplot2' library. geom_shadowpath(), geom_shadowline(), geom_shadowstep() and geom_shadowpoint() functions draw a shadow below lines to make busy plots more aesthetically pleasing. geom_glowpath(), geom_glowline(), geom_glowstep() and geom_glowpoint() add a neon glow around lines to get a steampunk style.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcmenem/ggshadow/}
  \item Report bugs at \url{https://github.com/marcmenem/ggshadow/issues}
}

}
\author{
\strong{Maintainer}: Marc Menem \email{marc.menem@m4x.org}

}
\keyword{internal}
