% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshadow-package.R
\docType{package}
\name{ggshadow-package}
\alias{ggshadow}
\alias{ggshadow-package}
\title{ggshadow: Shadow and Glow Geoms for 'ggplot2'}
\description{
A collection of Geoms for R's 'ggplot2' library. GeomShadowLine, GeomShadowPath, GeomShadowPoint, and GeomShadowStep support drawing a shadow below lines or around points to make busy plots more aesthetically pleasing. GeomGlowLine, GeomGlowPath, GeomGlowPoint, and GeomGlowStep support adding neon glow around lines or points to get a steampunk style.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcmenem/ggshadow/}
  \item Report bugs at \url{https://github.com/marcmenem/ggshadow/issues}
}

}
\author{
\strong{Maintainer}: Marc Menem \email{marc.menem@m4x.org}

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

}
\keyword{internal}
